@@ -20,15 +20,17 @@ import android.media.AudioDeviceInfo;
 import android.media.AudioManager;
 import android.os.Build;
 import android.preference.PreferenceManager;
+import android.support.annotation.Nullable;
 import android.util.Log;
 
 import com.xiaominfc.apprtc.util.AppRTCUtils;
 
-import org.webrtc.ThreadUtils;
-
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
+
+
+import org.webrtc.ThreadUtils;
 
 /**
  * AppRTCAudioManager manages all audio related parts of the AppRTC demo.
@@ -53,21 +55,23 @@ public class AppRTCAudioManager {
   }
 
   /** Selected audio device change event. */
-  public static interface AudioManagerEvents {
+  public interface AudioManagerEvents {
     // Callback fired once audio device is changed or list of available audio devices changed.
     void onAudioDeviceChanged(
         AudioDevice selectedAudioDevice, Set<AudioDevice> availableAudioDevices);
   }
 
   private final Context apprtcContext;
+  @Nullable
   private AudioManager audioManager;
 
+  @Nullable
   private AudioManagerEvents audioManagerEvents;
   private AudioManagerState amState;
   private int savedAudioMode = AudioManager.MODE_INVALID;
-  private boolean savedIsSpeakerPhoneOn = false;
-  private boolean savedIsMicrophoneMute = false;
-  private boolean hasWiredHeadset = false;
+  private boolean savedIsSpeakerPhoneOn;
+  private boolean savedIsMicrophoneMute;
+  private boolean hasWiredHeadset;
 
   // Default audio device; speaker phone for video calls or earpiece for audio
   // only calls.
@@ -93,19 +97,20 @@ public class AppRTCAudioManager {
   // relative to the view screen of a device and can therefore be used to
   // assist device switching (close to ear <=> use headset earpiece if
   // available, far from ear <=> use speaker phone).
-  private AppRTCProximitySensor proximitySensor = null;
+  @Nullable private AppRTCProximitySensor proximitySensor;
 
   // Handles all tasks related to Bluetooth headset devices.
   private final AppRTCBluetoothManager bluetoothManager;
 
   // Contains a list of available audio devices. A Set collection is used to
   // avoid duplicate elements.
-  private Set<AudioDevice> audioDevices = new HashSet<AudioDevice>();
+  private Set<AudioDevice> audioDevices = new HashSet<>();
 
   // Broadcast receiver for wired headset intent broadcasts.
   private BroadcastReceiver wiredHeadsetReceiver;
 
   // Callback method for changes in audio focus.
+  @Nullable
   private AudioManager.OnAudioFocusChangeListener audioFocusChangeListener;
 
   /**
@@ -153,7 +158,7 @@ public class AppRTCAudioManager {
       hasWiredHeadset = (state == STATE_PLUGGED);
       updateAudioDeviceState();
     }
-  };
+  }
 
   /** Construction. */
   static AppRTCAudioManager create(Context context) {
@@ -182,19 +187,17 @@ public class AppRTCAudioManager {
     // Create and initialize the proximity sensor.
     // Tablet devices (e.g. Nexus 7) does not support proximity sensors.
     // Note that, the sensor will not be active until start() has been called.
-    proximitySensor = AppRTCProximitySensor.create(context, new Runnable() {
-      // This method will be called each time a state change is detected.
-      // Example: user holds his hand over the device (closer than ~5 cm),
-      // or removes his hand from the device.
-      public void run() {
-        onProximitySensorChangedState();
-      }
-    });
+    proximitySensor = AppRTCProximitySensor.create(context,
+        // This method will be called each time a state change is detected.
+        // Example: user holds his hand over the device (closer than ~5 cm),
+        // or removes his hand from the device.
+        this ::onProximitySensorChangedState);
 
     Log.d(TAG, "defaultAudioDevice: " + defaultAudioDevice);
     AppRTCUtils.logDeviceInfo(TAG);
   }
 
+  @SuppressWarnings("deprecation") // TODO(henrika): audioManager.requestAudioFocus() is deprecated.
   public void start(AudioManagerEvents audioManagerEvents) {
     Log.d(TAG, "start");
     ThreadUtils.checkIsOnMainThread();
@@ -224,7 +227,7 @@ public class AppRTCAudioManager {
       // logging for now.
       @Override
       public void onAudioFocusChange(int focusChange) {
-        String typeOfChange = "AUDIOFOCUS_NOT_DEFINED";
+        final String typeOfChange;
         switch (focusChange) {
           case AudioManager.AUDIOFOCUS_GAIN:
             typeOfChange = "AUDIOFOCUS_GAIN";
@@ -292,6 +295,7 @@ public class AppRTCAudioManager {
     Log.d(TAG, "AudioManager started");
   }
 
+  @SuppressWarnings("deprecation") // TODO(henrika): audioManager.abandonAudioFocus() is deprecated.
   public void stop() {
     Log.d(TAG, "stop");
     ThreadUtils.checkIsOnMainThread();
@@ -387,7 +391,7 @@ public class AppRTCAudioManager {
   /** Returns current set of available/selectable audio devices. */
   public Set<AudioDevice> getAudioDevices() {
     ThreadUtils.checkIsOnMainThread();
-    return Collections.unmodifiableSet(new HashSet<AudioDevice>(audioDevices));
+    return Collections.unmodifiableSet(new HashSet<>(audioDevices));
   }
 
   /** Returns the currently selected audio device. */
@@ -559,7 +563,7 @@ public class AppRTCAudioManager {
     }
 
     // Update selected audio device.
-    AudioDevice newAudioDevice = selectedAudioDevice;
+    final AudioDevice newAudioDevice;
 
     if (bluetoothManager.getState() == AppRTCBluetoothManager.State.SCO_CONNECTED) {
       // If a Bluetooth is connected, then it should be used as output audio
