@@ -24,10 +24,14 @@ import android.media.AudioManager;
 import android.os.Handler;
 import android.os.Looper;
 import android.os.Process;
+import android.support.annotation.Nullable;
 import android.util.Log;
+
+import com.xiaominfc.apprtc.util.AppRTCUtils;
+
 import java.util.List;
 import java.util.Set;
-import com.xiaominfc.apprtc.util.AppRTCUtils;
+
 import org.webrtc.ThreadUtils;
 
 /**
@@ -64,14 +68,18 @@ public class AppRTCBluetoothManager {
 
   private final Context apprtcContext;
   private final AppRTCAudioManager apprtcAudioManager;
+  @Nullable
   private final AudioManager audioManager;
   private final Handler handler;
 
   int scoConnectionAttempts;
   private State bluetoothState;
   private final BluetoothProfile.ServiceListener bluetoothServiceListener;
+  @Nullable
   private BluetoothAdapter bluetoothAdapter;
+  @Nullable
   private BluetoothHeadset bluetoothHeadset;
+  @Nullable
   private BluetoothDevice bluetoothDevice;
   private final BroadcastReceiver bluetoothHeadsetReceiver;
 
@@ -187,7 +195,7 @@ public class AppRTCBluetoothManager {
       }
       Log.d(TAG, "onReceive done: BT state=" + bluetoothState);
     }
-  };
+  }
 
   /** Construction. */
   static AppRTCBluetoothManager create(Context context, AppRTCAudioManager audioManager) {
@@ -390,6 +398,7 @@ public class AppRTCBluetoothManager {
   /**
    * Stubs for test mocks.
    */
+  @Nullable
   protected AudioManager getAudioManager(Context context) {
     return (AudioManager) context.getSystemService(Context.AUDIO_SERVICE);
   }
