@@ -16,6 +16,7 @@ import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
 import android.os.Build;
+import android.support.annotation.Nullable;
 import android.util.Log;
 
 import com.xiaominfc.apprtc.util.AppRTCUtils;
@@ -41,8 +42,8 @@ public class AppRTCProximitySensor implements SensorEventListener {
 
   private final Runnable onSensorStateListener;
   private final SensorManager sensorManager;
-  private Sensor proximitySensor = null;
-  private boolean lastStateReportIsNear = false;
+  @Nullable private Sensor proximitySensor;
+  private boolean lastStateReportIsNear;
 
   /** Construction */
   static AppRTCProximitySensor create(Context context, Runnable sensorStateListener) {
