@@ -24,6 +24,7 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
+import android.support.annotation.Nullable;
 import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.View;
@@ -31,6 +32,11 @@ import android.view.Window;
 import android.view.WindowManager;
 import android.view.WindowManager.LayoutParams;
 import android.widget.Toast;
+import java.io.IOException;
+import java.lang.RuntimeException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
 
 import org.webrtc.Camera1Enumerator;
 import org.webrtc.Camera2Enumerator;
@@ -47,12 +53,8 @@ import org.webrtc.StatsReport;
 import org.webrtc.SurfaceViewRenderer;
 import org.webrtc.VideoCapturer;
 import org.webrtc.VideoFileRenderer;
-import org.webrtc.VideoRenderer;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
+import org.webrtc.VideoFrame;
+import org.webrtc.VideoSink;
 
 /**
  * Activity for peer connection call setup, call waiting
@@ -61,56 +63,58 @@ import java.util.Set;
 public class CallActivity extends Activity implements AppRTCClient.SignalingEvents,
                                                       PeerConnectionClient.PeerConnectionEvents,
                                                       CallFragment.OnCallEvents {
-  private static final String TAG = CallActivity.class.getSimpleName();
+  private static final String TAG = "CallRTCClient";
 
-  public static final String EXTRA_ROOMID = "org.appspot.apprtc.ROOMID";
-  public static final String EXTRA_URLPARAMETERS = "org.appspot.apprtc.URLPARAMETERS";
-  public static final String EXTRA_LOOPBACK = "org.appspot.apprtc.LOOPBACK";
-  public static final String EXTRA_VIDEO_CALL = "org.appspot.apprtc.VIDEO_CALL";
-  public static final String EXTRA_SCREENCAPTURE = "org.appspot.apprtc.SCREENCAPTURE";
-  public static final String EXTRA_CAMERA2 = "org.appspot.apprtc.CAMERA2";
-  public static final String EXTRA_VIDEO_WIDTH = "org.appspot.apprtc.VIDEO_WIDTH";
-  public static final String EXTRA_VIDEO_HEIGHT = "org.appspot.apprtc.VIDEO_HEIGHT";
-  public static final String EXTRA_VIDEO_FPS = "org.appspot.apprtc.VIDEO_FPS";
+  public static final String EXTRA_ROOMID = "com.xiaominfc.apprtc.ROOMID";
+  public static final String EXTRA_URLPARAMETERS = "com.xiaominfc.apprtc.URLPARAMETERS";
+  public static final String EXTRA_LOOPBACK = "com.xiaominfc.apprtc.LOOPBACK";
+  public static final String EXTRA_VIDEO_CALL = "com.xiaominfc.apprtc.VIDEO_CALL";
+  public static final String EXTRA_SCREENCAPTURE = "com.xiaominfc.apprtc.SCREENCAPTURE";
+  public static final String EXTRA_CAMERA2 = "com.xiaominfc.apprtc.CAMERA2";
+  public static final String EXTRA_VIDEO_WIDTH = "com.xiaominfc.apprtc.VIDEO_WIDTH";
+  public static final String EXTRA_VIDEO_HEIGHT = "com.xiaominfc.apprtc.VIDEO_HEIGHT";
+  public static final String EXTRA_VIDEO_FPS = "com.xiaominfc.apprtc.VIDEO_FPS";
   public static final String EXTRA_VIDEO_CAPTUREQUALITYSLIDER_ENABLED =
       "org.appsopt.apprtc.VIDEO_CAPTUREQUALITYSLIDER";
-  public static final String EXTRA_VIDEO_BITRATE = "org.appspot.apprtc.VIDEO_BITRATE";
-  public static final String EXTRA_VIDEOCODEC = "org.appspot.apprtc.VIDEOCODEC";
-  public static final String EXTRA_HWCODEC_ENABLED = "org.appspot.apprtc.HWCODEC";
-  public static final String EXTRA_CAPTURETOTEXTURE_ENABLED = "org.appspot.apprtc.CAPTURETOTEXTURE";
-  public static final String EXTRA_FLEXFEC_ENABLED = "org.appspot.apprtc.FLEXFEC";
-  public static final String EXTRA_AUDIO_BITRATE = "org.appspot.apprtc.AUDIO_BITRATE";
-  public static final String EXTRA_AUDIOCODEC = "org.appspot.apprtc.AUDIOCODEC";
+  public static final String EXTRA_VIDEO_BITRATE = "com.xiaominfc.apprtc.VIDEO_BITRATE";
+  public static final String EXTRA_VIDEOCODEC = "com.xiaominfc.apprtc.VIDEOCODEC";
+  public static final String EXTRA_HWCODEC_ENABLED = "com.xiaominfc.apprtc.HWCODEC";
+  public static final String EXTRA_CAPTURETOTEXTURE_ENABLED = "com.xiaominfc.apprtc.CAPTURETOTEXTURE";
+  public static final String EXTRA_FLEXFEC_ENABLED = "com.xiaominfc.apprtc.FLEXFEC";
+  public static final String EXTRA_AUDIO_BITRATE = "com.xiaominfc.apprtc.AUDIO_BITRATE";
+  public static final String EXTRA_AUDIOCODEC = "com.xiaominfc.apprtc.AUDIOCODEC";
   public static final String EXTRA_NOAUDIOPROCESSING_ENABLED =
-      "org.appspot.apprtc.NOAUDIOPROCESSING";
-  public static final String EXTRA_AECDUMP_ENABLED = "org.appspot.apprtc.AECDUMP";
-  public static final String EXTRA_OPENSLES_ENABLED = "org.appspot.apprtc.OPENSLES";
-  public static final String EXTRA_DISABLE_BUILT_IN_AEC = "org.appspot.apprtc.DISABLE_BUILT_IN_AEC";
-  public static final String EXTRA_DISABLE_BUILT_IN_AGC = "org.appspot.apprtc.DISABLE_BUILT_IN_AGC";
-  public static final String EXTRA_DISABLE_BUILT_IN_NS = "org.appspot.apprtc.DISABLE_BUILT_IN_NS";
-  public static final String EXTRA_ENABLE_LEVEL_CONTROL = "org.appspot.apprtc.ENABLE_LEVEL_CONTROL";
+      "com.xiaominfc.apprtc.NOAUDIOPROCESSING";
+  public static final String EXTRA_AECDUMP_ENABLED = "com.xiaominfc.apprtc.AECDUMP";
+  public static final String EXTRA_SAVE_INPUT_AUDIO_TO_FILE_ENABLED =
+      "com.xiaominfc.apprtc.SAVE_INPUT_AUDIO_TO_FILE";
+  public static final String EXTRA_OPENSLES_ENABLED = "com.xiaominfc.apprtc.OPENSLES";
+  public static final String EXTRA_DISABLE_BUILT_IN_AEC = "com.xiaominfc.apprtc.DISABLE_BUILT_IN_AEC";
+  public static final String EXTRA_DISABLE_BUILT_IN_AGC = "com.xiaominfc.apprtc.DISABLE_BUILT_IN_AGC";
+  public static final String EXTRA_DISABLE_BUILT_IN_NS = "com.xiaominfc.apprtc.DISABLE_BUILT_IN_NS";
   public static final String EXTRA_DISABLE_WEBRTC_AGC_AND_HPF =
-      "org.appspot.apprtc.DISABLE_WEBRTC_GAIN_CONTROL";
-  public static final String EXTRA_DISPLAY_HUD = "org.appspot.apprtc.DISPLAY_HUD";
-  public static final String EXTRA_TRACING = "org.appspot.apprtc.TRACING";
-  public static final String EXTRA_CMDLINE = "org.appspot.apprtc.CMDLINE";
-  public static final String EXTRA_RUNTIME = "org.appspot.apprtc.RUNTIME";
-  public static final String EXTRA_VIDEO_FILE_AS_CAMERA = "org.appspot.apprtc.VIDEO_FILE_AS_CAMERA";
+      "com.xiaominfc.apprtc.DISABLE_WEBRTC_GAIN_CONTROL";
+  public static final String EXTRA_DISPLAY_HUD = "com.xiaominfc.apprtc.DISPLAY_HUD";
+  public static final String EXTRA_TRACING = "com.xiaominfc.apprtc.TRACING";
+  public static final String EXTRA_CMDLINE = "com.xiaominfc.apprtc.CMDLINE";
+  public static final String EXTRA_RUNTIME = "com.xiaominfc.apprtc.RUNTIME";
+  public static final String EXTRA_VIDEO_FILE_AS_CAMERA = "com.xiaominfc.apprtc.VIDEO_FILE_AS_CAMERA";
   public static final String EXTRA_SAVE_REMOTE_VIDEO_TO_FILE =
-      "org.appspot.apprtc.SAVE_REMOTE_VIDEO_TO_FILE";
+      "com.xiaominfc.apprtc.SAVE_REMOTE_VIDEO_TO_FILE";
   public static final String EXTRA_SAVE_REMOTE_VIDEO_TO_FILE_WIDTH =
-      "org.appspot.apprtc.SAVE_REMOTE_VIDEO_TO_FILE_WIDTH";
+      "com.xiaominfc.apprtc.SAVE_REMOTE_VIDEO_TO_FILE_WIDTH";
   public static final String EXTRA_SAVE_REMOTE_VIDEO_TO_FILE_HEIGHT =
-      "org.appspot.apprtc.SAVE_REMOTE_VIDEO_TO_FILE_HEIGHT";
+      "com.xiaominfc.apprtc.SAVE_REMOTE_VIDEO_TO_FILE_HEIGHT";
   public static final String EXTRA_USE_VALUES_FROM_INTENT =
-      "org.appspot.apprtc.USE_VALUES_FROM_INTENT";
-  public static final String EXTRA_DATA_CHANNEL_ENABLED = "org.appspot.apprtc.DATA_CHANNEL_ENABLED";
-  public static final String EXTRA_ORDERED = "org.appspot.apprtc.ORDERED";
-  public static final String EXTRA_MAX_RETRANSMITS_MS = "org.appspot.apprtc.MAX_RETRANSMITS_MS";
-  public static final String EXTRA_MAX_RETRANSMITS = "org.appspot.apprtc.MAX_RETRANSMITS";
-  public static final String EXTRA_PROTOCOL = "org.appspot.apprtc.PROTOCOL";
-  public static final String EXTRA_NEGOTIATED = "org.appspot.apprtc.NEGOTIATED";
-  public static final String EXTRA_ID = "org.appspot.apprtc.ID";
+      "com.xiaominfc.apprtc.USE_VALUES_FROM_INTENT";
+  public static final String EXTRA_DATA_CHANNEL_ENABLED = "com.xiaominfc.apprtc.DATA_CHANNEL_ENABLED";
+  public static final String EXTRA_ORDERED = "com.xiaominfc.apprtc.ORDERED";
+  public static final String EXTRA_MAX_RETRANSMITS_MS = "com.xiaominfc.apprtc.MAX_RETRANSMITS_MS";
+  public static final String EXTRA_MAX_RETRANSMITS = "com.xiaominfc.apprtc.MAX_RETRANSMITS";
+  public static final String EXTRA_PROTOCOL = "com.xiaominfc.apprtc.PROTOCOL";
+  public static final String EXTRA_NEGOTIATED = "com.xiaominfc.apprtc.NEGOTIATED";
+  public static final String EXTRA_ID = "com.xiaominfc.apprtc.ID";
+  public static final String EXTRA_ENABLE_RTCEVENTLOG = "com.xiaominfc.apprtc.ENABLE_RTCEVENTLOG";
 
   private static final int CAPTURE_PERMISSION_REQUEST_CODE = 1;
 
@@ -121,48 +125,51 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
   // Peer connection statistics callback period in ms.
   private static final int STAT_CALLBACK_PERIOD = 1000;
 
-  private class ProxyRenderer implements VideoRenderer.Callbacks {
-    private VideoRenderer.Callbacks target;
+  private static class ProxyVideoSink implements VideoSink {
+    private VideoSink target;
 
-    synchronized public void renderFrame(VideoRenderer.I420Frame frame) {
+    @Override
+    synchronized public void onFrame(VideoFrame frame) {
       if (target == null) {
         Logging.d(TAG, "Dropping frame in proxy because target is null.");
-        VideoRenderer.renderFrameDone(frame);
         return;
       }
 
-      target.renderFrame(frame);
+      target.onFrame(frame);
     }
 
-    synchronized public void setTarget(VideoRenderer.Callbacks target) {
+    synchronized public void setTarget(VideoSink target) {
       this.target = target;
     }
   }
 
-  private final ProxyRenderer remoteProxyRenderer = new ProxyRenderer();
-  private final ProxyRenderer localProxyRenderer = new ProxyRenderer();
-  private PeerConnectionClient peerConnectionClient = null;
+  private final ProxyVideoSink remoteProxyRenderer = new ProxyVideoSink();
+  private final ProxyVideoSink localProxyVideoSink = new ProxyVideoSink();
+  @Nullable private PeerConnectionClient peerConnectionClient;
+  @Nullable
   private AppRTCClient appRtcClient;
+  @Nullable
   private AppRTCClient.SignalingParameters signalingParameters;
-  private AppRTCAudioManager audioManager = null;
-  private EglBase rootEglBase;
+  @Nullable private AppRTCAudioManager audioManager;
+  @Nullable
   private SurfaceViewRenderer pipRenderer;
+  @Nullable
   private SurfaceViewRenderer fullscreenRenderer;
+  @Nullable
   private VideoFileRenderer videoFileRenderer;
-  private final List<VideoRenderer.Callbacks> remoteRenderers =
-      new ArrayList<VideoRenderer.Callbacks>();
+  private final List<VideoSink> remoteSinks = new ArrayList<>();
   private Toast logToast;
   private boolean commandLineRun;
-  private int runTimeMs;
   private boolean activityRunning;
   private AppRTCClient.RoomConnectionParameters roomConnectionParameters;
+  @Nullable
   private PeerConnectionClient.PeerConnectionParameters peerConnectionParameters;
-  private boolean iceConnected;
+  private boolean connected;
   private boolean isError;
   private boolean callControlFragmentVisible = true;
-  private long callStartedTimeMs = 0;
+  private long callStartedTimeMs;
   private boolean micEnabled = true;
-  private boolean screencaptureEnabled = false;
+  private boolean screencaptureEnabled;
   private static Intent mediaProjectionPermissionResultData;
   private static int mediaProjectionPermissionResultCode;
   // True if local view is in the fullscreen renderer.
@@ -174,6 +181,9 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
   private CpuMonitor cpuMonitor;
 
   @Override
+  // TODO(bugs.webrtc.org/8580): LayoutParams.FLAG_TURN_SCREEN_ON and
+  // LayoutParams.FLAG_SHOW_WHEN_LOCKED are deprecated.
+  @SuppressWarnings("deprecation")
   public void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     Thread.setDefaultUncaughtExceptionHandler(new UnhandledExceptionHandler(this));
@@ -182,17 +192,16 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     // adding content.
     requestWindowFeature(Window.FEATURE_NO_TITLE);
     getWindow().addFlags(LayoutParams.FLAG_FULLSCREEN | LayoutParams.FLAG_KEEP_SCREEN_ON
-        | LayoutParams.FLAG_DISMISS_KEYGUARD | LayoutParams.FLAG_SHOW_WHEN_LOCKED
-        | LayoutParams.FLAG_TURN_SCREEN_ON);
+        | LayoutParams.FLAG_SHOW_WHEN_LOCKED | LayoutParams.FLAG_TURN_SCREEN_ON);
     getWindow().getDecorView().setSystemUiVisibility(getSystemUiVisibility());
     setContentView(R.layout.activity_call);
 
-    iceConnected = false;
+    connected = false;
     signalingParameters = null;
 
     // Create UI controls.
-    pipRenderer = (SurfaceViewRenderer) findViewById(R.id.pip_video_view);
-    fullscreenRenderer = (SurfaceViewRenderer) findViewById(R.id.fullscreen_video_view);
+    pipRenderer = findViewById(R.id.pip_video_view);
+    fullscreenRenderer = findViewById(R.id.fullscreen_video_view);
     callFragment = new CallFragment();
     hudFragment = new HudFragment();
 
@@ -213,13 +222,13 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     });
 
     fullscreenRenderer.setOnClickListener(listener);
-    remoteRenderers.add(remoteProxyRenderer);
+    remoteSinks.add(remoteProxyRenderer);
 
     final Intent intent = getIntent();
+    final EglBase eglBase = EglBase.create();
 
     // Create video renderers.
-    rootEglBase = EglBase.create();
-    pipRenderer.init(rootEglBase.getEglBaseContext(), null);
+    pipRenderer.init(eglBase.getEglBaseContext(), null);
     pipRenderer.setScalingType(ScalingType.SCALE_ASPECT_FIT);
     String saveRemoteVideoToFile = intent.getStringExtra(EXTRA_SAVE_REMOTE_VIDEO_TO_FILE);
 
@@ -229,19 +238,19 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
       int videoOutHeight = intent.getIntExtra(EXTRA_SAVE_REMOTE_VIDEO_TO_FILE_HEIGHT, 0);
       try {
         videoFileRenderer = new VideoFileRenderer(
-            saveRemoteVideoToFile, videoOutWidth, videoOutHeight, rootEglBase.getEglBaseContext());
-        remoteRenderers.add(videoFileRenderer);
+            saveRemoteVideoToFile, videoOutWidth, videoOutHeight, eglBase.getEglBaseContext());
+        remoteSinks.add(videoFileRenderer);
       } catch (IOException e) {
         throw new RuntimeException(
             "Failed to open video file for output: " + saveRemoteVideoToFile, e);
       }
     }
-    fullscreenRenderer.init(rootEglBase.getEglBaseContext(), null);
+    fullscreenRenderer.init(eglBase.getEglBaseContext(), null);
     fullscreenRenderer.setScalingType(ScalingType.SCALE_ASPECT_FILL);
 
     pipRenderer.setZOrderMediaOverlay(true);
     pipRenderer.setEnableHardwareScaler(true /* enabled */);
-    fullscreenRenderer.setEnableHardwareScaler(true /* enabled */);
+    fullscreenRenderer.setEnableHardwareScaler(false /* enabled */);
     // Start with local feed in fullscreen and swap it to the pip when the call is connected.
     setSwappedFeeds(true /* isSwappedFeeds */);
 
@@ -304,14 +313,15 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
             intent.getIntExtra(EXTRA_AUDIO_BITRATE, 0), intent.getStringExtra(EXTRA_AUDIOCODEC),
             intent.getBooleanExtra(EXTRA_NOAUDIOPROCESSING_ENABLED, false),
             intent.getBooleanExtra(EXTRA_AECDUMP_ENABLED, false),
+            intent.getBooleanExtra(EXTRA_SAVE_INPUT_AUDIO_TO_FILE_ENABLED, false),
             intent.getBooleanExtra(EXTRA_OPENSLES_ENABLED, false),
             intent.getBooleanExtra(EXTRA_DISABLE_BUILT_IN_AEC, false),
             intent.getBooleanExtra(EXTRA_DISABLE_BUILT_IN_AGC, false),
             intent.getBooleanExtra(EXTRA_DISABLE_BUILT_IN_NS, false),
-            intent.getBooleanExtra(EXTRA_ENABLE_LEVEL_CONTROL, false),
-            intent.getBooleanExtra(EXTRA_DISABLE_WEBRTC_AGC_AND_HPF, false), dataChannelParameters);
+            intent.getBooleanExtra(EXTRA_DISABLE_WEBRTC_AGC_AND_HPF, false),
+            intent.getBooleanExtra(EXTRA_ENABLE_RTCEVENTLOG, false), dataChannelParameters);
     commandLineRun = intent.getBooleanExtra(EXTRA_CMDLINE, false);
-    runTimeMs = intent.getIntExtra(EXTRA_RUNTIME, 0);
+    int runTimeMs = intent.getIntExtra(EXTRA_RUNTIME, 0);
 
     Log.d(TAG, "VIDEO_FILE: '" + intent.getStringExtra(EXTRA_VIDEO_FILE_AS_CAMERA) + "'");
 
@@ -329,8 +339,10 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
         new AppRTCClient.RoomConnectionParameters(roomUri.toString(), roomId, loopback, urlParameters);
 
     // Create CPU monitor
-    cpuMonitor = new CpuMonitor(this);
-    hudFragment.setCpuMonitor(cpuMonitor);
+    if (CpuMonitor.isSupported()) {
+      cpuMonitor = new CpuMonitor(this);
+      hudFragment.setCpuMonitor(cpuMonitor);
+    }
 
     // Send intent arguments to fragments.
     callFragment.setArguments(intent.getExtras());
@@ -351,16 +363,16 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
       }, runTimeMs);
     }
 
-    peerConnectionClient = PeerConnectionClient.getInstance();
+    // Create peer connection client.
+    peerConnectionClient = new PeerConnectionClient(
+        getApplicationContext(), eglBase, peerConnectionParameters, CallActivity.this);
+    PeerConnectionFactory.Options options = new PeerConnectionFactory.Options();
     if (loopback) {
-      PeerConnectionFactory.Options options = new PeerConnectionFactory.Options();
       options.networkIgnoreMask = 0;
-      peerConnectionClient.setPeerConnectionFactoryOptions(options);
     }
-    peerConnectionClient.createPeerConnectionFactory(
-        getApplicationContext(), peerConnectionParameters, CallActivity.this);
+    peerConnectionClient.createPeerConnectionFactory(options);
 
-    if (screencaptureEnabled && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+    if (screencaptureEnabled) {
       startScreenCapture();
     } else {
       startCall();
@@ -411,7 +423,7 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     return getIntent().getBooleanExtra(EXTRA_CAPTURETOTEXTURE_ENABLED, false);
   }
 
-  private VideoCapturer createCameraCapturer(CameraEnumerator enumerator) {
+  private @Nullable VideoCapturer createCameraCapturer(CameraEnumerator enumerator) {
     final String[] deviceNames = enumerator.getDeviceNames();
 
     // First, try to find front facing camera
@@ -444,7 +456,7 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
   }
 
   @TargetApi(21)
-  private VideoCapturer createScreenCapturer() {
+  private @Nullable VideoCapturer createScreenCapturer() {
     if (mediaProjectionPermissionResultCode != Activity.RESULT_OK) {
       reportError("User didn't give permission to capture the screen.");
       return null;
@@ -468,7 +480,9 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     if (peerConnectionClient != null && !screencaptureEnabled) {
       peerConnectionClient.stopVideoSource();
     }
-    cpuMonitor.pause();
+    if (cpuMonitor != null) {
+      cpuMonitor.pause();
+    }
   }
 
   @Override
@@ -479,7 +493,9 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     if (peerConnectionClient != null && !screencaptureEnabled) {
       peerConnectionClient.startVideoSource();
     }
-    cpuMonitor.resume();
+    if (cpuMonitor != null) {
+      cpuMonitor.resume();
+    }
   }
 
   @Override
@@ -490,7 +506,6 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
       logToast.cancel();
     }
     activityRunning = false;
-    rootEglBase.release();
     super.onDestroy();
   }
 
@@ -530,7 +545,7 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
 
   // Helper functions.
   private void toggleCallControlFragmentVisibility() {
-    if (!iceConnected || !callFragment.isAdded()) {
+    if (!connected || !callFragment.isAdded()) {
       return;
     }
     // Show/hide call control fragment
@@ -601,14 +616,10 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
   private void disconnect() {
     activityRunning = false;
     remoteProxyRenderer.setTarget(null);
-    localProxyRenderer.setTarget(null);
+    localProxyVideoSink.setTarget(null);
     if (appRtcClient != null) {
       appRtcClient.disconnectFromRoom();
       appRtcClient = null;
-    }
-    if (peerConnectionClient != null) {
-      peerConnectionClient.close();
-      peerConnectionClient = null;
     }
     if (pipRenderer != null) {
       pipRenderer.release();
@@ -622,11 +633,15 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
       fullscreenRenderer.release();
       fullscreenRenderer = null;
     }
+    if (peerConnectionClient != null) {
+      peerConnectionClient.close();
+      peerConnectionClient = null;
+    }
     if (audioManager != null) {
       audioManager.stop();
       audioManager = null;
     }
-    if (iceConnected && !isError) {
+    if (connected && !isError) {
       setResult(RESULT_OK);
     } else {
       setResult(RESULT_CANCELED);
@@ -678,8 +693,8 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     });
   }
 
-  private VideoCapturer createVideoCapturer() {
-    VideoCapturer videoCapturer = null;
+  private @Nullable VideoCapturer createVideoCapturer() {
+    final VideoCapturer videoCapturer;
     String videoFileAsCamera = getIntent().getStringExtra(EXTRA_VIDEO_FILE_AS_CAMERA);
     if (videoFileAsCamera != null) {
       try {
@@ -688,7 +703,7 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
         reportError("Failed to open video file for emulated camera");
         return null;
       }
-    } else if (screencaptureEnabled && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+    } else if (screencaptureEnabled) {
       return createScreenCapturer();
     } else if (useCamera2()) {
       if (!captureToTexture()) {
@@ -712,7 +727,7 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
   private void setSwappedFeeds(boolean isSwappedFeeds) {
     Logging.d(TAG, "setSwappedFeeds: " + isSwappedFeeds);
     this.isSwappedFeeds = isSwappedFeeds;
-    localProxyRenderer.setTarget(isSwappedFeeds ? fullscreenRenderer : pipRenderer);
+    localProxyVideoSink.setTarget(isSwappedFeeds ? fullscreenRenderer : pipRenderer);
     remoteProxyRenderer.setTarget(isSwappedFeeds ? pipRenderer : fullscreenRenderer);
     fullscreenRenderer.setMirror(isSwappedFeeds);
     pipRenderer.setMirror(!isSwappedFeeds);
@@ -730,8 +745,8 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     if (peerConnectionParameters.videoCallEnabled) {
       videoCapturer = createVideoCapturer();
     }
-    peerConnectionClient.createPeerConnection(rootEglBase.getEglBaseContext(), localProxyRenderer,
-        remoteRenderers, videoCapturer, signalingParameters);
+    peerConnectionClient.createPeerConnection(
+        localProxyVideoSink, remoteSinks, videoCapturer, signalingParameters);
 
     if (signalingParameters.initiator) {
       logAndToast("Creating OFFER...");
@@ -888,8 +903,6 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
       @Override
       public void run() {
         logAndToast("ICE connected, delay=" + delta + "ms");
-        iceConnected = true;
-        callConnected();
       }
     });
   }
@@ -900,7 +913,30 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
       @Override
       public void run() {
         logAndToast("ICE disconnected");
-        iceConnected = false;
+      }
+    });
+  }
+
+  @Override
+  public void onConnected() {
+    final long delta = System.currentTimeMillis() - callStartedTimeMs;
+    runOnUiThread(new Runnable() {
+      @Override
+      public void run() {
+        logAndToast("DTLS connected, delay=" + delta + "ms");
+        connected = true;
+        callConnected();
+      }
+    });
+  }
+
+  @Override
+  public void onDisconnected() {
+    runOnUiThread(new Runnable() {
+      @Override
+      public void run() {
+        logAndToast("DTLS disconnected");
+        connected = false;
         disconnect();
       }
     });
@@ -914,7 +950,7 @@ public class CallActivity extends Activity implements AppRTCClient.SignalingEven
     runOnUiThread(new Runnable() {
       @Override
       public void run() {
-        if (!isError && iceConnected) {
+        if (!isError && connected) {
           hudFragment.updateEncoderStatistics(reports);
         }
       }
