@@ -10,20 +10,18 @@
 
 package com.xiaominfc.apprtc;
 
+import android.support.annotation.Nullable;
 import android.util.Log;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.webrtc.IceCandidate;
-import org.webrtc.PeerConnection;
-import org.webrtc.SessionDescription;
-
-import java.util.LinkedList;
+import java.util.ArrayList;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.webrtc.IceCandidate;
+import org.webrtc.SessionDescription;
 
 /**
  * Implementation of AppRTCClient that uses direct TCP connection as the signaling channel.
@@ -53,6 +51,7 @@ public class DirectRTCClient implements AppRTCClient, TCPChannelClient.TCPChanne
 
   private final ExecutorService executor;
   private final SignalingEvents events;
+  @Nullable
   private TCPChannelClient tcpClient;
   private RoomConnectionParameters connectionParameters;
 
@@ -233,7 +232,7 @@ public class DirectRTCClient implements AppRTCClient, TCPChannelClient.TCPChanne
 
       SignalingParameters parameters = new SignalingParameters(
           // Ice servers are not needed for direct connections.
-          new LinkedList<PeerConnection.IceServer>(),
+          new ArrayList<>(),
           isServer, // Server side acts as the initiator on direct connections.
           null, // clientId
           null, // wssUrl
@@ -269,7 +268,7 @@ public class DirectRTCClient implements AppRTCClient, TCPChannelClient.TCPChanne
 
         SignalingParameters parameters = new SignalingParameters(
             // Ice servers are not needed for direct connections.
-            new LinkedList<PeerConnection.IceServer>(),
+            new ArrayList<>(),
             false, // This code will only be run on the client side. So, we are not the initiator.
             null, // clientId
             null, // wssUrl
