@@ -20,6 +20,7 @@ import android.widget.ImageButton;
 import android.widget.TextView;
 
 import org.webrtc.StatsReport;
+import com.xiaominfc.apprtc.R;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -28,7 +29,6 @@ import java.util.Map;
  * Fragment for HUD statistics display.
  */
 public class HudFragment extends Fragment {
-  private View controlView;
   private TextView encoderStatView;
   private TextView hudViewBwe;
   private TextView hudViewConnection;
@@ -43,15 +43,15 @@ public class HudFragment extends Fragment {
   @Override
   public View onCreateView(
       LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-    controlView = inflater.inflate(R.layout.fragment_hud, container, false);
+    View controlView = inflater.inflate(R.layout.fragment_hud, container, false);
 
     // Create UI controls.
-    encoderStatView = (TextView) controlView.findViewById(R.id.encoder_stat_call);
-    hudViewBwe = (TextView) controlView.findViewById(R.id.hud_stat_bwe);
-    hudViewConnection = (TextView) controlView.findViewById(R.id.hud_stat_connection);
-    hudViewVideoSend = (TextView) controlView.findViewById(R.id.hud_stat_video_send);
-    hudViewVideoRecv = (TextView) controlView.findViewById(R.id.hud_stat_video_recv);
-    toggleDebugButton = (ImageButton) controlView.findViewById(R.id.button_toggle_debug);
+    encoderStatView = controlView.findViewById(R.id.encoder_stat_call);
+    hudViewBwe = controlView.findViewById(R.id.hud_stat_bwe);
+    hudViewConnection = controlView.findViewById(R.id.hud_stat_connection);
+    hudViewVideoSend = controlView.findViewById(R.id.hud_stat_video_send);
+    hudViewVideoRecv = controlView.findViewById(R.id.hud_stat_video_recv);
+    toggleDebugButton = controlView.findViewById(R.id.button_toggle_debug);
 
     toggleDebugButton.setOnClickListener(new View.OnClickListener() {
       @Override
@@ -105,7 +105,7 @@ public class HudFragment extends Fragment {
   }
 
   private Map<String, String> getReportMap(StatsReport report) {
-    Map<String, String> reportMap = new HashMap<String, String>();
+    Map<String, String> reportMap = new HashMap<>();
     for (StatsReport.Value value : report.values) {
       reportMap.put(value.name, value.value);
     }
