@@ -16,9 +16,9 @@ import android.content.SharedPreferences.OnSharedPreferenceChangeListener;
 import android.os.Bundle;
 import android.preference.ListPreference;
 import android.preference.Preference;
-
 import org.webrtc.Camera2Enumerator;
-import org.webrtc.voiceengine.WebRtcAudioUtils;
+import org.webrtc.audio.JavaAudioDeviceModule;
+import com.xiaominfc.apprtc.R;
 
 /**
  * Settings activity for AppRTC.
@@ -43,17 +43,18 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
   private String keyPrefAudioCodec;
   private String keyprefNoAudioProcessing;
   private String keyprefAecDump;
+  private String keyprefEnableSaveInputAudioToFile;
   private String keyprefOpenSLES;
   private String keyprefDisableBuiltInAEC;
   private String keyprefDisableBuiltInAGC;
   private String keyprefDisableBuiltInNS;
-  private String keyprefEnableLevelControl;
   private String keyprefDisableWebRtcAGCAndHPF;
   private String keyprefSpeakerphone;
 
   private String keyPrefRoomServerUrl;
   private String keyPrefDisplayHud;
   private String keyPrefTracing;
+  private String keyprefEnabledRtcEventLog;
 
   private String keyprefEnableDataChannel;
   private String keyprefOrdered;
@@ -84,11 +85,12 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
     keyPrefAudioCodec = getString(R.string.pref_audiocodec_key);
     keyprefNoAudioProcessing = getString(R.string.pref_noaudioprocessing_key);
     keyprefAecDump = getString(R.string.pref_aecdump_key);
+    keyprefEnableSaveInputAudioToFile =
+        getString(R.string.pref_enable_save_input_audio_to_file_key);
     keyprefOpenSLES = getString(R.string.pref_opensles_key);
     keyprefDisableBuiltInAEC = getString(R.string.pref_disable_built_in_aec_key);
     keyprefDisableBuiltInAGC = getString(R.string.pref_disable_built_in_agc_key);
     keyprefDisableBuiltInNS = getString(R.string.pref_disable_built_in_ns_key);
-    keyprefEnableLevelControl = getString(R.string.pref_enable_level_control_key);
     keyprefDisableWebRtcAGCAndHPF = getString(R.string.pref_disable_webrtc_agc_and_hpf_key);
     keyprefSpeakerphone = getString(R.string.pref_speakerphone_key);
 
@@ -103,6 +105,7 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
     keyPrefRoomServerUrl = getString(R.string.pref_room_server_url_key);
     keyPrefDisplayHud = getString(R.string.pref_displayhud_key);
     keyPrefTracing = getString(R.string.pref_tracing_key);
+    keyprefEnabledRtcEventLog = getString(R.string.pref_enable_rtceventlog_key);
 
     // Display the fragment as the main content.
     settingsFragment = new SettingsFragment();
@@ -139,11 +142,11 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
     updateSummary(sharedPreferences, keyPrefAudioCodec);
     updateSummaryB(sharedPreferences, keyprefNoAudioProcessing);
     updateSummaryB(sharedPreferences, keyprefAecDump);
+    updateSummaryB(sharedPreferences, keyprefEnableSaveInputAudioToFile);
     updateSummaryB(sharedPreferences, keyprefOpenSLES);
     updateSummaryB(sharedPreferences, keyprefDisableBuiltInAEC);
     updateSummaryB(sharedPreferences, keyprefDisableBuiltInAGC);
     updateSummaryB(sharedPreferences, keyprefDisableBuiltInNS);
-    updateSummaryB(sharedPreferences, keyprefEnableLevelControl);
     updateSummaryB(sharedPreferences, keyprefDisableWebRtcAGCAndHPF);
     updateSummaryList(sharedPreferences, keyprefSpeakerphone);
 
@@ -159,6 +162,7 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
     updateSummary(sharedPreferences, keyPrefRoomServerUrl);
     updateSummaryB(sharedPreferences, keyPrefDisplayHud);
     updateSummaryB(sharedPreferences, keyPrefTracing);
+    updateSummaryB(sharedPreferences, keyprefEnabledRtcEventLog);
 
     if (!Camera2Enumerator.isSupported(this)) {
       Preference camera2Preference = settingsFragment.findPreference(keyprefCamera2);
@@ -167,10 +171,7 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
       camera2Preference.setEnabled(false);
     }
 
-    // Disable forcing WebRTC based AEC so it won't affect our value.
-    // Otherwise, if it was enabled, isAcousticEchoCancelerSupported would always return false.
-    WebRtcAudioUtils.setWebRtcBasedAcousticEchoCanceler(false);
-    if (!WebRtcAudioUtils.isAcousticEchoCancelerSupported()) {
+    if (!JavaAudioDeviceModule.isBuiltInAcousticEchoCancelerSupported()) {
       Preference disableBuiltInAECPreference =
           settingsFragment.findPreference(keyprefDisableBuiltInAEC);
 
@@ -178,17 +179,13 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
       disableBuiltInAECPreference.setEnabled(false);
     }
 
-    WebRtcAudioUtils.setWebRtcBasedAutomaticGainControl(false);
-    if (!WebRtcAudioUtils.isAutomaticGainControlSupported()) {
-      Preference disableBuiltInAGCPreference =
-          settingsFragment.findPreference(keyprefDisableBuiltInAGC);
+    Preference disableBuiltInAGCPreference =
+        settingsFragment.findPreference(keyprefDisableBuiltInAGC);
 
-      disableBuiltInAGCPreference.setSummary(getString(R.string.pref_built_in_agc_not_available));
-      disableBuiltInAGCPreference.setEnabled(false);
-    }
+    disableBuiltInAGCPreference.setSummary(getString(R.string.pref_built_in_agc_not_available));
+    disableBuiltInAGCPreference.setEnabled(false);
 
-    WebRtcAudioUtils.setWebRtcBasedNoiseSuppressor(false);
-    if (!WebRtcAudioUtils.isNoiseSuppressorSupported()) {
+    if (!JavaAudioDeviceModule.isBuiltInNoiseSuppressorSupported()) {
       Preference disableBuiltInNSPreference =
           settingsFragment.findPreference(keyprefDisableBuiltInNS);
 
@@ -233,16 +230,17 @@ public class SettingsActivity extends Activity implements OnSharedPreferenceChan
         || key.equals(keyprefFlexfec)
         || key.equals(keyprefNoAudioProcessing)
         || key.equals(keyprefAecDump)
+        || key.equals(keyprefEnableSaveInputAudioToFile)
         || key.equals(keyprefOpenSLES)
         || key.equals(keyprefDisableBuiltInAEC)
         || key.equals(keyprefDisableBuiltInAGC)
         || key.equals(keyprefDisableBuiltInNS)
-        || key.equals(keyprefEnableLevelControl)
         || key.equals(keyprefDisableWebRtcAGCAndHPF)
         || key.equals(keyPrefDisplayHud)
         || key.equals(keyprefEnableDataChannel)
         || key.equals(keyprefOrdered)
-        || key.equals(keyprefNegotiated)) {
+        || key.equals(keyprefNegotiated)
+        || key.equals(keyprefEnabledRtcEventLog)) {
       updateSummaryB(sharedPreferences, key);
     } else if (key.equals(keyprefSpeakerphone)) {
       updateSummaryList(sharedPreferences, key);
