@@ -12,10 +12,12 @@ package com.xiaominfc.apprtc;
 
 import android.os.Handler;
 import android.os.HandlerThread;
+import android.support.annotation.Nullable;
 import android.util.Log;
 
 import com.xiaominfc.apprtc.util.AsyncHttpURLConnection;
-import com.xiaominfc.apprtc.util.AsyncHttpURLConnection.AsyncHttpEvents;
+import com.xiaominfc.apprtc.WebSocketChannelClient.WebSocketChannelEvents;
+import com.xiaominfc.apprtc.WebSocketChannelClient.WebSocketConnectionState;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -33,7 +35,7 @@ import org.webrtc.SessionDescription;
  * Messages to other party (with local Ice candidates and answer SDP) can
  * be sent after WebSocket connection is established.
  */
-public class WebSocketRTCClient implements AppRTCClient, WebSocketChannelClient.WebSocketChannelEvents {
+public class WebSocketRTCClient implements AppRTCClient, WebSocketChannelEvents {
   private static final String TAG = "WSRTCClient";
   private static final String ROOM_JOIN = "join";
   private static final String ROOM_MESSAGE = "message";
@@ -288,7 +290,7 @@ public class WebSocketRTCClient implements AppRTCClient, WebSocketChannelClient.
   // (passed to WebSocket client constructor).
   @Override
   public void onWebSocketMessage(final String msg) {
-    if (wsClient.getState() != WebSocketChannelClient.WebSocketConnectionState.REGISTERED) {
+    if (wsClient.getState() != WebSocketConnectionState.REGISTERED) {
       Log.e(TAG, "Got WebSocket message in non registered state.");
       return;
     }
@@ -377,14 +379,14 @@ public class WebSocketRTCClient implements AppRTCClient, WebSocketChannelClient.
 
   // Send SDP or ICE candidate to a room server.
   private void sendPostMessage(
-      final MessageType messageType, final String url, final String message) {
+      final MessageType messageType, final String url, @Nullable final String message) {
     String logInfo = url;
     if (message != null) {
       logInfo += ". Message: " + message;
     }
     Log.d(TAG, "C->GAE: " + logInfo);
     AsyncHttpURLConnection httpConnection =
-        new AsyncHttpURLConnection("POST", url, message, new AsyncHttpEvents() {
+        new AsyncHttpURLConnection("POST", url, message, new AsyncHttpURLConnection.AsyncHttpEvents() {
           @Override
           public void onHttpError(String errorMessage) {
             reportError("GAE POST error: " + errorMessage);
@@ -418,7 +420,7 @@ public class WebSocketRTCClient implements AppRTCClient, WebSocketChannelClient.
   }
 
   // Converts a JSON candidate to a Java object.
-  private IceCandidate toJavaCandidate(JSONObject json) throws JSONException {
+  IceCandidate toJavaCandidate(JSONObject json) throws JSONException {
     return new IceCandidate(
         json.getString("id"), json.getInt("label"), json.getString("candidate"));
   }
